deleted unneeded tests folders, moved firecrawl to connectors/external/
